requirements.txt
numpy >= 1.24.1
pandas >= 1.5.3
scipy >= 1.10.0
dash >= 2.8.1
plotly >= 5.12.0
dash-bootstrap-component >= 1.3.1